pyproject.toml
[project]
name = "auditor-expert"
version = "0.1.0"
description = "RAG-powered ISO 9001 / IATF 16949 / AS9100 audit expert"
requires-python = ">=3.11"

dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "chromadb>=0.5.0",
    "langchain-text-splitters>=0.3.0",
    "tiktoken>=0.8.0",
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",
    "gradio>=4.40.0",
    "langfuse>=2.50.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "matplotlib>=3.9.0",
    "plotly>=5.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.black]
line-length = 100
target-version = ["py311"]